+++ src/askForPromise.js
@@ -3,6 +3,9 @@
    askForPromise Description
    ========================
    Returns object with promise and resolve function.
+   - Created March 12th, 2016;
+   - Promise with timeout added July 16th, 2017 (v.1.3.0);
+   - askForPromise.all & AskForPromise.sequence added October 15th, 2023(v.1.4.0); 
 
 */
 
@@ -26,6 +29,53 @@ function askForPromise ( list ) {
    askObject.timeout = _timeout ( isList, askObject )   
    return askObject
  } // askForPromise func.
+
+
+
+
+
+ askForPromise.sequence = function promiseInSequence ( list, ...args ) {
+  const 
+        task = askForPromise ()
+      , result = []
+      ;
+
+  function* listGen ( n ) {   for ( const el of n ) { yield el }} 
+  const g = listGen ( list );
+
+  function wait ( n, ...args ) {
+      if ( n.done ) {
+               task.done ( result )
+               return
+          }
+      n.value (...args).then ( r => {
+              result.push ( r )
+              wait( g.next(), ...args, r )
+          }) 
+      } // wait func.
+
+  wait ( g.next(), ...args )
+  return task
+} // promiseInSequence func.
+
+
+
+
+
+askForPromise.all = function promiseAll ( list, ...args ) {
+  const 
+        task = askForPromise ()
+      , result = []
+      , r = list.map ( (n,i) => { 
+                            return (typeof n === 'function') ? n(...args).then ( r => result[i] = r   ) 
+                                                             : n.then ( r => result[i] = r   )
+                        })
+      ;
+  Promise.all ( r ).then ( () => task.done(result)   )
+  return task
+} // promiseAll func.
+
+
 
 
 
@@ -57,10 +107,10 @@ function _singlePromise () {
 
 
 
- function _after (x) {
-    return function ( fx ) {
+function _after (x) {
+return function ( fx ) {
                 x.then ( res => fx(res)   )
-       }}
+}}